fix: change bare range upper bound from unbounded to bounded
NOTE: the maven docs imply that '1.0' should be expanded to '[1.0,)'
(an unbounded upper range, as per the previous implementation here)
but that causes issues:

For example: '3.2.9' then "satisfies" '3.2.6', and during upgrade
path calculation this results in it looking like transitive
dependencies are merely out of date, and a `mvn install` would fix
the issue, when this is not the case.

Also, `maven install` seems not to follow this rule itself, and
specifying '1.0' pulls in '1.0' even if a newer version exists.