feat: consider unknown invalid
We have started to set versions to 'unknown' during Maven version resolution.
These versions can appear when the resolver cannot get a valid version.
For example if the version is set in a parent, but the parent is inacessible (private repository or download failed).

Changing maven-semver to consider 'unknown' version input as invalid.
This will prevent false positives - vulnerabilities on packages with
'unknown' versions.